Refactoed the folder structure and completed the swagger definiton for the summoner
